Moved UICommandEntry to the org.terasology.nui.widgets package.
