+++ engine/src/main/java/org/destinationsol/ui/nui/widgets/UICommandEntry.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2014 MovingBlocks
+ * Copyright 2020 The Terasology Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *  http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.destinationsol.ui.nui.screens;
+package org.destinationsol.ui.nui.widgets;
 
 import com.google.common.collect.Lists;
 import org.destinationsol.game.console.TabCompletionEngine;
@@ -26,6 +26,7 @@ import org.terasology.nui.widgets.UIText;
 import java.util.List;
 
 /**
+ * A UI Widget used for command-line input and tab-completion in the console screen.
  */
 public class UICommandEntry extends UIText {
     private Binding<List<String>> commandHistory = new DefaultBinding<>(Lists.<String>newArrayList());
